update scratch deploy (interim)
- deploy insurance contracts
- deploy el rewards contract
- start protocol in dao-local-deploy
- add aux script vote-and-enact
- refactor
