+++ scripts/multisig/11-finalize-dao.js
@@ -65,7 +65,7 @@ async function finalizeDAO({ web3, artifacts }) {
 
   log.splitter()
 
-  await saveCallTxData(`finalizeDAO`, template, 'finalizeDAO', `tx-07-finalize-dao.json`, {
+  await saveCallTxData(`finalizeDAO`, template, 'finalizeDAO', `tx-11-finalize-dao.json`, {
     arguments: [
       state.daoAragonId,
       percentToBP(fee.totalPercent),
